Update README.md
In this project  a virtual reality paradigm written in python language that allows was developed to measure the behavioral responses (avoidance and escape) of voluntaries facing with moderate (low fear) and intense threat (high fear). Concomitant to the behavioral responses the  autonomic reactivity (respiratory and heart rates) are also measured to detect the emotional susceptibility of the individuals to the stressful conditions. The virtual paradigm consists of a familiarization phase, control and low and high fear phases of 1 min each, in which the collision between predator and prey represented by geometrical symbols (red and blue rectangles, respectively) produces a bell tone (control-collA) or a computer beep if the prey goes outside the open field (escA). In the next two phases, collision produces a white noise at 80 or 85 dB  (low and high fear conditions, respectively). Collisiion or escape occurring during these phases are measured as coll B and esc B.